+++ week4/forecast/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import PageLayout from "../components/PageLayout";
+import PageLayout from "../components/common/PageLayout";
 
 const Home = () => {
   // 기본 홈이 되는 페이지
